refactor: move account code into shared folder
