+++ src/account/components/AccountProfile.js
@@ -15,6 +15,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import React, { useCallback } from 'react';
+
 import _ from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
@@ -22,10 +23,12 @@ import {
   fetchProfile,
   savePreference,
   saveUser,
-} from 'terrasoApi/account/accountSlice';
+} from 'terrasoApi/shared/account/accountSlice';
 import { useFetchData } from 'terrasoApi/utils';
 import * as yup from 'yup';
+
 import { Checkbox, FormControlLabel } from '@mui/material';
+
 import { useDocumentDescription, useDocumentTitle } from 'common/document';
 import Form from 'forms/components/Form';
 import PageContainer from 'layout/PageContainer';
@@ -33,6 +36,7 @@ import PageHeader from 'layout/PageHeader';
 import PageLoader from 'layout/PageLoader';
 import LocalePickerSelect from 'localization/components/LocalePickerSelect';
 import { useAnalytics } from 'monitoring/analytics';
+
 import AccountAvatar from './AccountAvatar';
 
 const VALIDATION_SCHEMA = yup
