+++ src/account/components/RequireAuth.js
@@ -17,7 +17,7 @@
 import React, { useCallback } from 'react';
 import { useSelector } from 'react-redux';
 import { Navigate, useLocation } from 'react-router-dom';
-import { fetchUser } from 'terrasoApi/account/accountSlice';
+import { fetchUser } from 'terrasoApi/shared/account/accountSlice';
 import { useFetchData } from 'terrasoApi/utils';
 import PageLoader from 'layout/PageLoader';
 import { generateReferrerPath } from 'navigation/navigationUtils';
