+++ src/account/components/Unsubscribe.js
@@ -15,13 +15,16 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import { useCallback, useEffect } from 'react';
+
 import _ from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { unsubscribeFromNotifications } from 'terrasoApi/account/accountSlice';
+import { unsubscribeFromNotifications } from 'terrasoApi/shared/account/accountSlice';
 import { useFetchData } from 'terrasoApi/utils';
+
 import { Alert } from '@mui/material';
+
 import { useDocumentTitle } from 'common/document';
 import PageContainer from 'layout/PageContainer';
 import PageLoader from 'layout/PageLoader';
