+++ src/localization/components/LocalePicker.js
@@ -15,10 +15,12 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import React, { useEffect } from 'react';
+
 import _ from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
-import { savePreference } from 'terrasoApi/account/accountSlice';
+import { savePreference } from 'terrasoApi/shared/account/accountSlice';
+
 import LocalePickerSelect from 'localization/components/LocalePickerSelect';
 import { LOCALES } from 'localization/i18n';
