+++ src/terrasoApi/shared/account/auth.ts
@@ -16,7 +16,8 @@
  */
 import Cookies from 'js-cookie';
 import jwt from 'jwt-decode';
-import { UNAUTHENTICATED } from 'terrasoApi/account/authConstants';
+import { UNAUTHENTICATED } from 'terrasoApi/shared/account/authConstants';
+
 import { COOKIES_DOMAIN, TERRASO_API_URL } from 'config';
 
 const COOKIES_PARAMS = { path: '/', domain: COOKIES_DOMAIN };
