+++ src/terrasoApi/terrasoBackend/api.ts
@@ -15,10 +15,12 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import _ from 'lodash/fp';
-import { getAuthHeaders } from 'terrasoApi/account/auth';
-import { UNAUTHENTICATED } from 'terrasoApi/account/authConstants';
 import { TypedDocumentString } from 'terrasoApi/gql/graphql';
+import { getAuthHeaders } from 'terrasoApi/shared/account/auth';
+import { UNAUTHENTICATED } from 'terrasoApi/shared/account/authConstants';
+
 import logger from 'monitoring/logger';
+
 import { GRAPH_QL_ENDPOINT, TERRASO_API_URL } from 'config';
 
 type Error = { message: any };
